requirements.txt
streamlit
pandas
plotly
fpdf
redis
python-dotenv
supabase
